fixup! chore(readonly-boolean): add new private component
