+++ src/components/label/examples/label.scss
@@ -1,4 +1,4 @@
-:host(limel-example-readonly-boolean) {
+:host(limel-example-label) {
     --example-controls-column-layout: auto-fit;
     display: grid;
     gap: 1rem;
